refactor(functions): extract find candidates
