+++ test/functions_test.js
@@ -21,6 +21,7 @@ import {
   anywayGet,
   filterMove,
   buildMove,
+  findCandidates,
 } from "../src/functions.js";
 
 {
@@ -784,5 +785,182 @@ import {
   ];
   (async () => {
     assert.deepEqual(await buildMove(input), expected);
+  })();
+}
+
+{
+  /** @type {import("../src/models").BoardPanel[][]} */
+  const board = [
+    ["a", "b"],
+    [null, null],
+  ];
+  /** @type {import("../src/models").Coordinate[]} */
+  const coordinates = [];
+  (async () => {
+    assert.deepEqual(await findCandidates(board, coordinates), [null, null]);
+  })();
+}
+
+{
+  /** @type {import("../src/models").BoardPanel[][]} */
+  const board = [
+    ["a", null],
+    ["a", null],
+  ];
+  /** @type {import("../src/models").Coordinate[]} */
+  const coordinates = [{ x: 1, y: 0, panel: "x" }];
+  const expected = [null, ["ax"]];
+  (async () => {
+    assert.deepEqual(await findCandidates(board, coordinates), expected);
+  })();
+}
+
+{
+  /** @type {import("../src/models").BoardPanel[][]} */
+  const board = [
+    ["a", null],
+    ["a", null],
+  ];
+  /** @type {import("../src/models").Coordinate[]} */
+  const coordinates = [{ x: 1, y: 1, panel: "x" }];
+  const expected = [null, ["ax"]];
+  (async () => {
+    assert.deepEqual(await findCandidates(board, coordinates), expected);
+  })();
+}
+
+{
+  /** @type {import("../src/models").BoardPanel[][]} */
+  const board = [
+    ["a", null, "b"],
+    ["a", null, null],
+  ];
+  /** @type {import("../src/models").Coordinate[]} */
+  const coordinates = [{ x: 1, y: 0, panel: "x" }];
+  const expected = [null, ["axb"]];
+  (async () => {
+    assert.deepEqual(await findCandidates(board, coordinates), expected);
+  })();
+}
+
+{
+  /** @type {import("../src/models").BoardPanel[][]} */
+  const board = [
+    ["a", null, "b", null],
+    ["a", null, null, null],
+  ];
+  /** @type {import("../src/models").Coordinate[]} */
+  const coordinates = [
+    { x: 1, y: 0, panel: "x" },
+    { x: 3, y: 0, panel: "y" },
+  ];
+  const expected = [null, ["axby"]];
+  (async () => {
+    assert.deepEqual(await findCandidates(board, coordinates), expected);
+  })();
+}
+
+{
+  /** @type {import("../src/models").BoardPanel[][]} */
+  const board = [
+    [null, null, "b", null],
+    ["a", null, null, null],
+  ];
+  /** @type {import("../src/models").Coordinate[]} */
+  const coordinates = [
+    { x: 1, y: 0, panel: "x" },
+    { x: 3, y: 0, panel: "y" },
+  ];
+  const expected = [null, ["xby"]];
+  (async () => {
+    assert.deepEqual(await findCandidates(board, coordinates), expected);
+  })();
+}
+
+{
+  /** @type {import("../src/models").BoardPanel[][]} */
+  const board = [
+    ["a", null],
+    [null, null],
+  ];
+  /** @type {import("../src/models").Coordinate[]} */
+  const coordinates = [{ x: 0, y: 1, panel: "x" }];
+  const expected = [null, ["ax"]];
+  (async () => {
+    assert.deepEqual(await findCandidates(board, coordinates), expected);
+  })();
+}
+
+{
+  /** @type {import("../src/models").BoardPanel[][]} */
+  const board = [
+    ["a", null],
+    [null, null],
+    ["a", null],
+  ];
+  /** @type {import("../src/models").Coordinate[]} */
+  const coordinates = [{ x: 0, y: 1, panel: "x" }];
+  const expected = [null, ["axa"]];
+  (async () => {
+    assert.deepEqual(await findCandidates(board, coordinates), expected);
+  })();
+}
+
+{
+  /** @type {import("../src/models").BoardPanel[][]} */
+  const board = [
+    ["a", null],
+    [null, null],
+    ["a", null],
+    [null, null],
+  ];
+  /** @type {import("../src/models").Coordinate[]} */
+  const coordinates = [
+    { x: 0, y: 1, panel: "x" },
+    { x: 0, y: 3, panel: "x" },
+  ];
+  const expected = [null, ["axax"]];
+  (async () => {
+    assert.deepEqual(await findCandidates(board, coordinates), expected);
+  })();
+}
+
+{
+  /** @type {import("../src/models").BoardPanel[][]} */
+  const board = [
+    ["a", null],
+    [null, null],
+  ];
+  /** @type {import("../src/models").Coordinate[]} */
+  const coordinates = [
+    { x: 0, y: 1, panel: "x" },
+    { x: 1, y: 1, panel: "y" },
+  ];
+  const expected = [null, ["xy", "ax"]];
+  (async () => {
+    assert.deepEqual(await findCandidates(board, coordinates), expected);
+  })();
+}
+
+{
+  /** @type {import("../src/models").BoardPanel[][]} */
+  const board = [
+    ["a", null],
+    [null, null],
+    [null, null],
+  ];
+  /** @type {import("../src/models").Coordinate[]} */
+  const coordinates = [
+    { x: 1, y: 0, panel: "x" },
+    { x: 1, y: 2, panel: "y" },
+  ];
+  (async () => {
+    const [errors] = await findCandidates(board, coordinates);
+    if (errors !== null) {
+      assert.equal(errors.length, 1);
+      assert.equal(errors[0].message, "Empty panel x: 1, y: 1");
+    } else {
+      assert.fail("unreachable");
+    }
   })();
 }
